+++ semantic-conventions/CHANGELOG.md
@@ -16,6 +16,10 @@ Please update the changelog as part of any significant pull request.
   ([#271](https://github.com/open-telemetry/build-tools/pull/271))
 - Add link to requirement levels definition from Markdown table title.
   ([#222](https://github.com/open-telemetry/build-tools/pull/222))
+- Added code-generation mode that groups attributes by the root namespace and ability to write each group into individual file.
+  [BREAKING] The `--file-per-group <pattern>` that used to create multiple directories (like `output/<pattern>/file`) now generates
+  multiple files (`output/<pattern>file`) instead.
+  ([#243](https://github.com/open-telemetry/build-tools/pull/243))
 
 ## v0.23.0